move utils funcs to standalone file
